Update README.md
In the process of trying to build the app, I ran across a couple things. I originally assumed Node 10.0.0 was old since it's the very first release, so I went with Node 12 LTS, only to find that Realm didn't support it. Calling out the latest version of 10 seemed more secure to me. 

I suppose most developers will be using MacOS since it's required for building iOS, but I ran into the "security" command missing. It fails in a cryptic way too.