421 rtl text ios (#492)
* WIP: adding RTL support for text in iOS

In iOS <Text> components do not show RTL text by default. You need to either
add alignSelf to the style of the <Text> component, or re-style the
container. Here we've chosen to replace uses <Text> with our own <StyledText>
that explicitly adds alignSelf so that text in RTL languages always displays
in RTL in iOS.

This is a WIP because we have not replaced uses of <Text> with <StyledText>
everwhere.

* Use textAlign left instead of alignSelf start

* Replace current usage of Text with StyledText

Co-authored-by: Ken-ichi Ueda <kenichi.ueda@gmail.com>