Replace toast showing code
I have removed a boolean for the match screen that checks if the screen is rendered for the first time or more. I did so because it was keeping the challenge toast not appearing for the second or more observations that counted towards the challenge.
I have replaced the logic of not showing the badge or challenge toast when coming back to the match screen, from species details for example, with a state that tracks which toasts were already shown.