Fix state bug on initial camera roll load (#716)
* Fix state bug on initial camera roll load

This issue has been bothering me for years and it took me a very long
time to find the time to debug it with Xcode on my own local device.

1. load Seek
2. tap the camera icon
3. tap the "PHOTOS" tab at the bottom right
4. a new PHOTO LIBRARY page shows a spinning wheel which eventually
  resolves to a blank page of a slightly different color
5. if tap the "PHOTO LIBRARY" title and switch to any other album,
photos eventually do load and get populated in the gallery, and this
also works switching back to "PHOTO LIBRARY" (the default), hinting that
this is a state bug.

It seems like @jtklein was at least aware of it based on this
comment from last year in the [cameraRollHelpers checkForUniquePhotos
helper function][1].

Mutating the `seen` state within the body of a helper function is not a
good idea because this state mutation is invisible to React.

Having `checkForUniquePhotos()` return a tuple with both the
`uniqAssets` array and the `newSeen` set allows us to call the `dispatch`
method from `useReducer` to set the state of `seen` to the new state
we get from `newSeen`.

Locally I also updated the @react-native-camera-roll/camera-roll
dependency to the more recent 7.1.0 version which supports Smart Albums
so let me know if you'd be interested to handle that.

[1]: https://github.com/olivierlacan/SeekReactNative/commit/1fd9e1508353504a2437dc2fca028093f5060394

* Remove previous idea of tackling this issue

* Some code styling

* Change return type key

* Sorry for being pedantic

But now it is both alphabetical and in symmetry with the function input params.

* Migrate camerRollHelpers to TypeScript

---------

Co-authored-by: Johannes Klein <johannes.t.klein@gmail.com>