Do no longer save photos to a Seek album
Our fork of react-native-camera-roll returns a placeholder uri if we have Add-only permission and not Read (because we can not return the actual uri without read permission).

If this is the case we use the temporary photo file for the resizing steps and so forth that happen after photo is taken.