Enable react compiler (#820)
* Add react compiler dev dependencies

* Enable eslint plugin

* Add react compiler babel plugin

* Update package-lock.json

* Like this the app runs

And I can see automatic memoization working in the Dev Tools.
If I leave it in I get an error "can't read ObservationProvider of undefined".
So, branch is WIP

* Does not need to be initialized as undefined.

* Update prop type

* This ternary was thorwing an error with react compiler

Not exactly sure why, but the code change should be with the same result. In case the chained access is erroring out the catch block will return null.

* Update frame processor log

* Ignore lint

* Remove unnecessary dispatch call

* Update package.json